Drag n drop can be used for different users

+++ db/delete_from_layout.sql
@@ -1,1 +1,1 @@
-DELETE FROM profile_layout WHERE pen_id = $1;
+DELETE FROM profile_layout WHERE pen_id = $1 AND user_id = $2;
